+++ package.json
@@ -18,7 +18,7 @@
     "dev:node:api": "ts-node -r dotenv/config src/index.ts",
     "format": "pretty-quick",
     "lint": "eslint \"src/**/*.ts\" --fix",
-    "test": "node --expose-gc --no-compilation-cache ./node_modules/.bin/jest --detectOpenHandles --runInBand --logHeapUsage --detectLeaks",
+    "test": "cross-env NODE_ENV=test jest --detectOpenHandles --runInBand --logHeapUsage --detectLeaks",
     "test:upd": "yarn test -u",
     "test:path": "yarn test --no-cache --testPathPattern",
     "test:clean": "jest --clearCache",
@@ -30,20 +30,20 @@
   "dependencies": {
     "@aws/dynamodb-data-mapper": "^0.7.3",
     "@aws/dynamodb-data-mapper-annotations": "^0.7.3",
-    "@badger-dao/sdk": "^1.17.0",
-    "@tsed/common": "^6.95.9",
-    "@tsed/core": "^6.95.9",
-    "@tsed/di": "^6.95.9",
-    "@tsed/exceptions": "^6.95.9",
-    "@tsed/json-mapper": "^6.95.9",
-    "@tsed/platform-aws": "^6.95.9",
-    "@tsed/platform-express": "^6.95.9",
-    "@tsed/schema": "^6.95.9",
-    "@tsed/swagger": "^6.95.9",
+    "@badger-dao/sdk": "^1.17.1",
+    "@tsed/common": "^6.126.1",
+    "@tsed/core": "^6.126.1",
+    "@tsed/di": "^6.126.1",
+    "@tsed/exceptions": "^6.126.1",
+    "@tsed/json-mapper": "^6.126.1",
+    "@tsed/platform-aws": "^6.102.8-rc.1",
+    "@tsed/platform-express": "^6.126.1",
+    "@tsed/schema": "^6.126.1",
+    "@tsed/swagger": "^6.126.1",
     "@types/cors": "^2.8.12",
     "aws-sdk": "^2.1045.0",
     "axios": "^0.24.0",
-    "body-parser": "^1.19.0",
+    "body-parser": "^1.20.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.13.2",
     "cookie-parser": "^1.4.6",
@@ -58,7 +58,7 @@
     "tslib": "^2.2.0"
   },
   "devDependencies": {
-    "@badger-dao/sdk-mocks": "^1.16.11",
+    "@badger-dao/sdk-mocks": "^1.17.1",
     "@graphql-codegen/cli": "^2.3.0",
     "@graphql-codegen/typescript": "^2.7.3",
     "@graphql-codegen/typescript-graphql-request": "^4.5.3",
@@ -77,6 +77,7 @@
     "@typescript-eslint/eslint-plugin": "^5.33.0",
     "@typescript-eslint/parser": "^5.33.0",
     "clean-modules": "^2.0.4",
+    "cross-env": "^7.0.3",
     "eslint": "^8.21.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.26.0",
@@ -96,7 +97,7 @@
     "serverless-dynamodb-local": "^0.2.39",
     "serverless-offline": "^8.2.0",
     "serverless-plugin-typescript": "^2.1.2",
-    "supertest": "6.1.6",
+    "supertest": "6.2.4",
     "ts-jest": "^28.0.7",
     "ts-node": "^10.0.0",
     "typechain": "^7.0.0",
