+++ src/test/tests.utils.ts
@@ -26,15 +26,12 @@ import { BinanceSmartChain } from '../chains/config/bsc.config';
 import { Ethereum } from '../chains/config/eth.config';
 import { Fantom } from '../chains/config/fantom.config';
 import { Polygon } from '../chains/config/polygon.config';
-import { ChainVaults } from '../chains/vaults/chain.vaults';
 import { LeaderBoardType } from '../leaderboards/enums/leaderboard-type.enum';
 import * as pricesUtils from '../prices/prices.utils';
 import { fullTokenMockMap } from '../tokens/mocks/full-token.mock';
 import { historicVaultSnapshotsMock } from '../vaults/mocks/historic-vault-snapshots.mock';
 import { vaultsChartDataMock } from '../vaults/mocks/vaults-chart-data.mock';
 import { MOCK_VAULT_DEFINITION, TEST_ADDR, TEST_CURRENT_BLOCK } from './constants';
-// import { Chain } from '../chains/config/chain.config';
-// import { TestEthereum } from '../chains/config/teth.config';
 
 export function setupMapper(items: unknown[], filter?: (items: unknown[]) => unknown[]) {
   // @ts-ignore
@@ -287,11 +284,6 @@ export function mockPricing() {
     }
     return price / 2;
   });
-}
-
-export function mockChainVaults() {
-  jest.spyOn(ChainVaults.prototype, 'getVault').mockImplementation(async (_) => MOCK_VAULT_DEFINITION);
-  jest.spyOn(ChainVaults.prototype, 'all').mockImplementation(async () => [MOCK_VAULT_DEFINITION]);
 }
 
 export async function mockBadgerSdk(
